% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{format_for_slda}
\alias{format_for_slda}
\title{Summarize data frame}
\usage{
format_for_slda(df, counts, dilutions, cutoff, groupby = NULL)
}
\arguments{
\item{counts}{Column name of counts data}

\item{cutoff}{Threshold to score positive wells}

\item{groupby}{Optional sample identifier}

\item{dilution}{Column name of dilutions tested}
}
\value{
A data.frame object summarizing results of LDA
}
\description{
Helper function to summarize data for input to \link[SLDAssay:get.mle]{SLDAssay::get.mle}
}
