% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{format_results_as_df}
\alias{format_results_as_df}
\title{Convert list to data frame}
\usage{
format_results_as_df(results)
}
\arguments{
\item{results}{List returned from \link[SLDAssay:get.mle]{SLDAssay::get.mle}}
}
\value{
A data.frame of results
}
\description{
Helper function to format list output from \link[SLDAssay:get.mle]{SLDAssay::get.mle} into a data frame
}
