% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{getMLE}
\alias{getMLE}
\title{Get Maximum Likelihood Estimate (MLE)}
\usage{
getMLE(
  df,
  counts,
  dilutions,
  groupby = NULL,
  cutoff = 25,
  conf_level = 0.95,
  mc = 15000
)
}
\arguments{
\item{df}{A data.frame object containing columns 'counts' and 'dilutions' or a path to a .csv file containing those columns}

\item{counts}{Character name of column in df where cell counts are located}

\item{dilutions}{Character name of column in df where cell dilutions are located}

\item{groupby}{Optional character name of column in df containing sample or group names. The MLE will be calculated for each sample in this column. If not provided, will treat the entire data frame as one sample (the default)}

\item{cutoff}{Threshold for scoring counts as positive or negative (default: 25 cells)}

\item{conf_level}{The confidence level of the interval (default: 0.95)}

\item{mc}{Number of Monte Carlo samples. Default is exact (no MC sampling), unless more than 15,000 possible positive well outcomes exist, in which case 15,000 MC samples are taken. Use mc=FALSE for exact computation}
}
\value{
Results as a list, where \code{results$summary} is the data frame summarizing the LDA experiment, and \code{results$statistics} is the maximum likelihood estimate, confidence interval, and goodness of fit p-value
}
\description{
Calculates the MLE for LDA experiment(s). Uses \link[SLDAssay:get.mle]{SLDAssay::get.mle} under the hood.
}
\details{
Returned statistics include:
\itemize{
\item \code{MLE}: MLE
\item \code{BC_MLE}: Bias-corrected MLE
\item \code{Exact_PGOF}: P-value for goodness of fit. The probability of an experimental result as rare as or rarer than that obtained, assuming that the model is correct. Low values (e.g. < 0.01) indicate rare or implausible experimental results. Samples with a very low values might be considered for retesting
\item \code{Asymp_PGOF}: P-value calculated using an asymptotic Chi-Squared distribution with D-1 degrees of freedom, where D is the number of dilution levels
\item \code{Exact_CI}: Exact confidence interval, computed from the likelihood ratio test (recommended)
\item \code{Asymp_CI}: Wald asymptotic confidence interval, based on the normal approximation to the binomial distribution
}
}
