% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_one}
\alias{plot_one}
\title{Plot one sample and return. Argument \code{stats} must be one row.
Helper function for \code{\link[=plotLDA]{plotLDA()}}. No input validation done here.}
\usage{
plot_one(df, stats, dilutions, groupby = NULL, signf_dig = 3)
}
\arguments{
\item{df}{The data.frame from results$summary returned from \code{\link[=getMLE]{getMLE()}}}

\item{stats}{The data.frame from results$statistics returned from \code{\link[=getMLE]{getMLE()}}}

\item{dilutions}{Column name in df and stats}

\item{groupby}{Column name in df and stats or NULL}

\item{signf_dig}{Number of significant digits to use for statistics label}
}
\value{
A ggplot
}
\description{
Plot one sample and return. Argument \code{stats} must be one row.
Helper function for \code{\link[=plotLDA]{plotLDA()}}. No input validation done here.
}
