% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_theme}
\alias{plot_theme}
\title{A custom ggplot2 theme builder.}
\usage{
plot_theme(
  font_size = 10,
  font_family = "sans",
  line_width = 0.5,
  legend_position = "none",
  top_margin = 1,
  bottom_margin = 0.25,
  right_margin = 0.1,
  left_margin = 0.25,
  margin_units = "cm",
  device_font = "win"
)
}
\arguments{
\item{font_size}{Font size (default: 10)}

\item{font_family}{Font family (default: "sans")}

\item{line_width}{Line width (default: 0.5)}

\item{legend_position}{Legend position (default: "none")}

\item{top_margin}{Top margin (default: 1 cm)}

\item{bottom_margin}{Bottom margin (default: 0.25 cm)}

\item{right_margin}{Right margin (default: 0.1 cm)}

\item{left_margin}{Left margin (default: 0.25 cm)}

\item{margin_units}{Margin units (default: "cm")}

\item{device_font}{Loads font libraries from the users computer. See the \link[extrafont:loadfonts]{extrafont::loadfonts} documentation for details}
}
\value{
A ggplot2 theme object
}
\description{
Can be added to new \code{ggplot2} plots using \code{p <- p + plot_theme()}. Further
customization is possible by applying a new \link[ggplot2:theme]{ggplot2::theme} to the returned plot
(see \code{?ggplot2::theme()} for details).
}
