% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotLDA}
\alias{plotLDA}
\title{Plot results of LDA}
\usage{
plotLDA(
  results,
  dilutions,
  groupby = NULL,
  signf_dig = 2,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{results}{Output from \code{\link[=getMLE]{getMLE()}}}

\item{dilutions}{Column name of dilutions. Should be as provided to \code{\link[=getMLE]{getMLE()}}}

\item{groupby}{Optional character name of column containing sample or group names. Should be as provided to \code{\link[=getMLE]{getMLE()}}}

\item{signf_dig}{Number of significant digits to include in statistics (default: 2)}

\item{nrow}{Number of rows to plot. Only used when argument 'groupby' provided.}

\item{ncol}{Number of columns to plot. Only used when argument 'groupby' provided}

\item{...}{Additional arguments passed to \code{\link[=plot_theme]{plot_theme()}}}
}
\value{
A single ggplot or grid of ggplots
}
\description{
Plots dilutions along the x-axis and \% negative wells along the y-axis.
Adds a label including the bias-corrected MLE, exact CI, exact GoF p-value,
and cutoff threshold.
}
