% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMLE}
\alias{plotMLE}
\title{Plot MLE with confidence intervals}
\usage{
plotMLE(
  results,
  groupby,
  width = 0.5,
  mle_colour = "black",
  ci_colour = "gray",
  ...
)
}
\arguments{
\item{results}{Output from \code{\link[=getMLE]{getMLE()}}}

\item{groupby}{Character name of column containing sample or group names. Should be as provided to \code{\link[=getMLE]{getMLE()}}}

\item{mle_colour}{Line colour of the MLE value (default: "black")}

\item{ci_colour}{Line colour of the CI box (default: "gray")}

\item{...}{Additional arguments passed to \code{\link[=plot_theme]{plot_theme()}}}
}
\value{
A ggplot object
}
\description{
Creates a plot showing the bias-corrected MLE and the upper and lower
exact confidence interval for each sample in the column specified by \code{groupby}.
}
